database table header metadata should from table result as optimize table now should has zero dependency
