+++ src/components/board/board-chart-editor.tsx
@@ -12,8 +12,8 @@ import { ChartValue } from "../chart/chart-type";
 import EditChartMenu from "../chart/edit-chart-menu";
 import ResultTable from "../gui/query-result-table";
 import SqlEditor from "../gui/sql-editor";
-import { TableHeaderMetadata } from "../gui/table-optimized";
 import OptimizeTableState from "../gui/table-optimized/optimize-table-state";
+import { TableHeaderMetadata } from "../gui/table-result/type";
 import { Button } from "../orbit/button";
 import { MenuBar } from "../orbit/menu-bar";
 import { createAutoBoardChartValue } from "./board-auto-value";
