+++ src/components/gui/table-result/context-menu.tsx
@@ -10,9 +10,9 @@ import { KEY_BINDING } from "@/lib/key-matcher";
 import { LucidePlus, LucideTrash2 } from "lucide-react";
 import { useCallback } from "react";
 import { useFullEditor } from "../providers/full-editor-provider";
-import { TableHeaderMetadata } from "../table-optimized";
 import OptimizeTableState from "../table-optimized/optimize-table-state";
 import TableStateActions from "../table-optimized/table-state-actions";
+import { TableHeaderMetadata } from "./type";
 
 export default function useTableResultContextMenu({
   tableName,
